added attention score manipulation test

+++ backend/tests/util.py
@@ -12,7 +12,7 @@ def checkAlphaPixels(image, ref_image,alpha):
     for x in range(width):
         for y in range(height):
             for c in range(channels):
-                if image[x][y][c] != ref_image[x][y][c]* alpha:
+                if image[x][y][c] != int(ref_image[x][y][c]* alpha):
                     print(image[x][y][c])
                     print(ref_image[x][y][c]* alpha)
                     print(f"X{x} y{y} c{c}")
